mime type svg

+++ docker/index.js
@@ -28,6 +28,9 @@ http.createServer(function (request, response) {
 					return response.end(JSON.stringify({ err: { errorMessage: 'deleteTable forbidden in demo'}, }));
 				if ( event._POST.method === 'updateTable' && (demo_tables.indexOf(event._POST.payload.TableName) !== -1) )
 					return response.end(JSON.stringify({ err: { errorMessage: 'updateTable forbidden in demo'}, }));
+
+				if ( event._POST.method === 'putItem'  && (demo_tables.indexOf(event._POST.payload.TableName) !== -1) )
+					return response.end(JSON.stringify({ err: { errorMessage: 'putItem forbidden in demo'}, }));
 				if ( event._POST.method === 'deleteItem'  && (demo_tables.indexOf(event._POST.payload.TableName) !== -1) )
 					return response.end(JSON.stringify({ err: { errorMessage: 'deleteItem forbidden in demo'}, }));
 
@@ -43,6 +46,7 @@ http.createServer(function (request, response) {
 				case 'scan':
 				case 'query':
 				case 'deleteItem':
+				case 'putItem':
 					dynamodb[event._POST.method](event._POST.payload, function(err, data) {
 						response.end(JSON.stringify({ err: err, data:data }));
 					})
